requirements/garch.txt
pandas
numpy
scipy
statsmodels
arch
fredapi
pyarrow
pyyaml
tabulate
matplotlib
